flow/requirements.txt
# LangChain Core
langchain
langchain-core

# OpenAI Integration
langchain-openai
openai

# Anthropic Integration
langchain-anthropic

# Google Gemini (PaLM) Integration
langchain-google-genai
google-generativeai

# Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

# Environment Variable Management
python-dotenv

# FastAPI & Server
fastapi
uvicorn[standard]
python-multipart
gunicorn

# Groq Integration
langchain-groq

# YouTube Transcript
youtube-transcript-api

# Embeddings (local, no API token needed)
sentence-transformers

# Vector Store
langchain-community
faiss-cpu